+++ upgrade/upgrade-1.0.12.php
@@ -1,0 +1,7 @@
+<?php
+if (!defined('_PS_VERSION_')) { exit; }
+
+function upgrade_module_1_0_12($module): bool
+{
+    return true;
+}